clean up, reorganize
